requirements.txt
transformers
torch
streamlit
google-generativeai
python-dotenv